refactor(appointment): 약속 잡기 생성 폼 내 컴포넌트 추출
- 동일한 페이지에서 서로 달라질 확률이 거의 없어 보일 때, 하나의 컴포넌트를 재사용하여 UI를 구성해주었습니다.
